+++ src/test/components/spreadAny/spreadAny.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+function Component({ children, ...props }: any) {
+  return (
+    <div {...props}>
+      <div>{children}</div>
+    </div>
+  );
+}
+